backend/requirements.txt
agno>=2.0.0
fastapi[standard]>=0.112.0
sqlalchemy>=2.0.30
pydantic>=2.7
python-dotenv>=1.0.1
mcp>=1.2.0
openai>=1.40.0
anthropic>=0.30.0
uvicorn>=0.30.0
